Refactoring again to fix codacy issues
